Finish merge with structural refactoring. Get rid of cyclic dependencies
