+++ src/dishka/container.py
@@ -15,13 +15,13 @@ from .container_objects import Exit
 from .context_proxy import ContextProxy
 from .dependency_source import Factory
 from .entities.validation_settigs import DEFAULT_VALIDATION, ValidationSettings
-from .exceptions.base import ExitError
-from .exceptions.fabrication import NoFactoryError
-from .exceptions.scopes import (
+from .exceptions import (
     ChildScopeNotFoundError,
     NoChildScopesError,
     NoNonSkippedScopesError,
+    NoFactoryError
 )
+from .exception_base import ExitError
 from .provider import BaseProvider
 from .registry import Registry
 from .registry_builder import RegistryBuilder
