--- src/dishka/exceptions/__init__.py
@@ -1,31 +1,0 @@
-from dishka.exceptions.base import (
-    DishkaError,
-    ExitError,
-    NoContextValueError,
-)
-from dishka.exceptions.fabrication import (
-    NoFactoryError,
-    UnsupportedFactoryError,
-)
-from dishka.exceptions.graph import (
-    CycleDependenciesError,
-    GraphMissingFactoryError,
-    ImplicitOverrideDetectedError,
-    InvalidGraphError,
-    NothingOverriddenError,
-    UnknownScopeError,
-)
-
-__all__ = [
-    "CycleDependenciesError",
-    "DishkaError",
-    "ExitError",
-    "GraphMissingFactoryError",
-    "ImplicitOverrideDetectedError",
-    "InvalidGraphError",
-    "NoContextValueError",
-    "NoFactoryError",
-    "NothingOverriddenError",
-    "UnknownScopeError",
-    "UnsupportedFactoryError",
-]
