+++ src/dishka/provider/exceptions.py
@@ -1,17 +1,16 @@
-# noqa: A005
 from collections.abc import Sequence
-from typing import Any, TypeVar
+from typing import Any
 
-from dishka.exceptions.base import DishkaError
+from dishka.exception_base import DishkaError
 from dishka.text_rendering import get_name
 
 
-class UnsupportedGenericBoundsError(TypeError, DishkaError):
-    def __init__(self, bounds: TypeVar) -> None:
-        self.bounds = bounds
+class NotAFactoryError(TypeError, DishkaError):
+    def __init__(self, attempted_factory_type: type) -> None:
+        self.type = attempted_factory_type
 
     def __str__(self) -> str:
-        return f"Generic bounds are not supported: {self.bounds}"
+        return f"Cannot use {self.type} as a factory"
 
 
 class UnsupportedGeneratorReturnTypeError(TypeError, DishkaError):
@@ -94,4 +93,39 @@ class MissingReturnHintError(ValueError, DishkaError):
         return (
             f"Failed to analyze `{name}`. \n"
             f"Missing return type hint."
+        )
+
+
+class NoScopeSetInProvideError(ValueError, DishkaError):
+    def __init__(
+            self,
+            provides_name: str,
+            src_name: str,
+            provider_name: str,
+    ) -> None:
+        self.provides_name = provides_name
+        self.src_name = src_name
+        self.provider_name = provider_name
+
+    def __str__(self) -> str:
+        return (
+            f"No scope is set for {self.provides_name}.\n"
+            f"Set in provide() call for {self.src_name} or "
+            f"within {self.provider_name}"
+        )
+
+
+class NoScopeSetInContextError(ValueError, DishkaError):
+    def __init__(
+            self,
+            provides_name: str,
+            provider_name: str,
+    ) -> None:
+        self.provides_name = provides_name
+        self.provider_name = provider_name
+
+    def __str__(self) -> str:
+        return (
+            f"No scope is set for {self.provides_name}.\n"
+            f"Set in from_context() call or within {self.provider_name}"
         )
