+++ tests/unit/container/test_alias.py
@@ -14,7 +14,7 @@ from dishka import (
     make_container,
     provide,
 )
-from dishka.exceptions.graph import CycleDependenciesError, NoFactoryError
+from dishka.exceptions import CycleDependenciesError, NoFactoryError
 
 
 class AliasProvider(Provider):
